+++ core/src/main/java/com/contentful/vault/Vault.java
@@ -102,8 +102,12 @@ public class Vault {
         .build());
   }
 
-  public <T extends Resource> Query<T> fetch(Class<T> type) {
-    return new Query<T>(type, this);
+  public <T extends Resource> FetchQuery<T> fetch(Class<T> type) {
+    return new FetchQuery<T>(type, this);
+  }
+
+  public <T extends Resource> ObserveQuery<T> observe(Class<T> type) {
+    return new ObserveQuery<T>(type, this);
   }
 
   public SQLiteDatabase getReadableDatabase() {
@@ -128,16 +132,6 @@ public class Vault {
       }
       SQLITE_HELPERS.clear();
     }
-  }
-
-  private <T extends Resource> ModelHelper<?> getModelHelperOrThrow(SpaceHelper spaceHelper,
-      Class<T> clazz) {
-    ModelHelper<?> modelHelper = spaceHelper.getModels().get(clazz);
-    if (modelHelper == null) {
-      throw new IllegalArgumentException(
-          "Unable to find table mapping for class \"" + clazz.getName() + "\".");
-    }
-    return modelHelper;
   }
 
   private static SqliteHelper createSqliteHelper(Context context, SpaceHelper spaceHelper) {